refactored searchDatasets resolver
