+++ packages/openneuro-app/src/scripts/search/search-results.tsx
@@ -1,88 +1,8 @@
 import React from 'react'
-import { gql, useQuery } from '@apollo/client'
+import { useQuery } from '@apollo/client'
 import { useParams } from 'react-router-dom'
-
+import { datasets } from '@openneuro/client'
 import { datasetQueryDisplay } from '../datalad/dashboard/datasets/dataset-query.jsx'
-
-const searchQuery = gql`
-  query searchDatasets($q: String!, $cursor: String) {
-    datasets: search(q: $q, first: 25, after: $cursor) {
-      edges {
-        node {
-          id
-          created
-          uploader {
-            id
-            name
-          }
-          public
-          permissions {
-            id
-            userPermissions {
-              userId
-              level
-              access: level
-              user {
-                id
-                name
-                email
-                provider
-              }
-            }
-          }
-          draft {
-            id
-            summary {
-              modalities
-              sessions
-              subjects
-              subjectMetadata {
-                participantId
-                age
-                sex
-                group
-              }
-              tasks
-              size
-              totalFiles
-              dataProcessed
-            }
-            issues {
-              severity
-            }
-            description {
-              Name
-            }
-          }
-          analytics {
-            views
-            downloads
-          }
-          stars {
-            userId
-            datasetId
-          }
-          followers {
-            userId
-            datasetId
-          }
-          snapshots {
-            id
-            created
-            tag
-          }
-        }
-      }
-      pageInfo {
-        startCursor
-        endCursor
-        hasPreviousPage
-        hasNextPage
-        count
-      }
-    }
-  }
-`
 
 /**
  * This component is responsible for obtaining results from Elastic based
@@ -94,7 +14,7 @@ const SearchResultsQuery: React.FC = () => {
     true,
     false,
   )(
-    useQuery(searchQuery, {
+    useQuery(datasets.searchDatasets, {
       variables: {
         q: query,
       },
