+++ packages/openneuro-client/src/datasets.js
@@ -181,6 +181,86 @@ export const getDatasets = gql`
   }
 `
 
+export const searchDatasets = gql`
+  query searchDatasets($q: String!, $cursor: String) {
+    searchDatasets(q: $q, first: 25, after: $cursor) {
+      edges {
+        node {
+          id
+          created
+          uploader {
+            id
+            name
+          }
+          public
+          permissions {
+            id
+            userPermissions {
+              userId
+              level
+              access: level
+              user {
+                id
+                name
+                email
+                provider
+              }
+            }
+          }
+          draft {
+            id
+            summary {
+              modalities
+              sessions
+              subjects
+              subjectMetadata {
+                participantId
+                age
+                sex
+                group
+              }
+              tasks
+              size
+              totalFiles
+              dataProcessed
+            }
+            issues {
+              severity
+            }
+            description {
+              Name
+            }
+          }
+          analytics {
+            views
+            downloads
+          }
+          stars {
+            userId
+            datasetId
+          }
+          followers {
+            userId
+            datasetId
+          }
+          snapshots {
+            id
+            created
+            tag
+          }
+        }
+      }
+      pageInfo {
+        startCursor
+        endCursor
+        hasPreviousPage
+        hasNextPage
+        count
+      }
+    }
+  }
+`
+
 export const getDatasetIssues = gql`
   query getDatasetIssues($datasetId: ID!) {
     dataset(id: $datasetId) {
