+++ packages/openneuro-server/src/graphql/resolvers/query.js
@@ -6,10 +6,12 @@ import { snapshot, participantCount } from './snapshots.js'
 import { user, users } from './user.js'
 import { datasetChanges } from './dataset-change.js'
 import { flaggedFiles } from './flaggedFiles'
+import { searchDatasets } from './dataset-search'
 
 const Query = {
   dataset,
   datasets,
+  searchDatasets,
   user,
   users,
   snapshot,
